[WIP] move navigation.metrics package into the internal package
