+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/NavigationTelemetry.java
@@ -17,14 +17,14 @@ import com.mapbox.geojson.utils.PolylineUtils;
 import com.mapbox.services.android.navigation.BuildConfig;
 import com.mapbox.services.android.navigation.v5.internal.exception.NavigationException;
 import com.mapbox.services.android.navigation.v5.internal.location.MetricsLocation;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.FeedbackEvent;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationMetricListener;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.RerouteEvent;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.SessionState;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.TelemetryEvent;
+import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress;
 import com.mapbox.services.android.navigation.v5.navigation.MapboxNavigation;
 import com.mapbox.services.android.navigation.v5.navigation.MapboxNavigationOptions;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.FeedbackEvent;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.NavigationMetricListener;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.RerouteEvent;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.SessionState;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.TelemetryEvent;
-import com.mapbox.services.android.navigation.v5.routeprogress.MetricsRouteProgress;
 import com.mapbox.services.android.navigation.v5.routeprogress.RouteProgress;
 import com.mapbox.services.android.navigation.v5.utils.RingBuffer;
